+++ R/SeverityEstimate-package.R
@@ -1,0 +1,8 @@
+#' @keywords internal
+#' @import methods
+#' @import Rcpp
+#' @importFrom RcppParallel RcppParallelLibs
+#' @importFrom rstan sampling
+#' @importFrom rstantools rstan_config
+#' @useDynLib SeverityEstimate, .registration = TRUE
+"_PACKAGE"